requirements.txt
langchain~=0.1.20
langchain-community
langchain-openai
python-dotenv~=1.0.1
requests~=2.31.0